[C++] Renamed repo and added pdf files.
Okay, so I renamed the repository and added pdf files for the different
tasks. This was done to archive the tasks, if I want to work on them in
the feature.

Some of the tasks also have code that didn't score 100% on the site.
These files will be improved at a later time.